добавил валидацию на объект локальный администратор

+++ src/Application/Actions/LocalAdmin/LocalAdminAction.php
@@ -2,7 +2,10 @@
 namespace App\Application\Actions\LocalAdmin;
 
 use App\Application\Actions\Action;
+use App\Application\Middleware\AuthorizeMiddleware;
 use App\Services\LocalAdmin\LocalAdminService;
+use App\Validators\LocalAdmin\LocalAdminValidator;
+use App\Validators\Organization\OrganizationObjectValidator;
 use Psr\Http\Message\RequestInterface as Request;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ResponseInterface as Response;
@@ -18,8 +21,24 @@ class LocalAdminAction extends Action
 
     public function addWithoutMessageToEmail(Request $request, Response $response, $args):Response
     {
-        //TODO валидация на то, что эту операцию пытается делать глобальный администратор
+        if ($this->tokenWithError($response, $request)){
+            return $response->withStatus(401);
+        }
+        $userRole = $request->getHeader('userRole')[0];
+
+        if ($userRole != AuthorizeMiddleware::GLOBAL_ADMIN){
+            return $response->withStatus(403);
+        }
+
         $rowParams = json_decode($request->getBody()->getContents(), true);
+        $rowParams['organizationId'] = (int)$args['id'];
+        $rowParams['localAdminId'] = -1;
+
+        $errors = (new LocalAdminValidator())->validate($rowParams);
+        if (count($errors) > 0){
+            return $this->respond(400, ['errors' => $errors], $response);
+        }
+
         //TODO валидация объекта локальный администратор
         $localAdminId = $this->localAdminService->addWithoutMessage($rowParams['name'], $rowParams['password'], $rowParams['email'], (int)$args['id'], $response);
 
@@ -31,7 +50,15 @@ class LocalAdminAction extends Action
 
     public function delete(Request $request, Response $response, $args)
     {
-        //TODO валидация на то, что эту операцию пытается делать глобальный администратор
+        if ($this->tokenWithError($response, $request)){
+            return $response->withStatus(401);
+        }
+
+        $userRole = $request->getHeader('userRole')[0];
+        if ($userRole != AuthorizeMiddleware::GLOBAL_ADMIN){
+            return $response->withStatus(403);
+        }
+
         $idOrganization = (int)$args['id'];
         $idLocalAdmin = (int)$args['idLocalAdmin'];
         $this->localAdminService->delete($idLocalAdmin, $idOrganization, $response);
@@ -40,6 +67,15 @@ class LocalAdminAction extends Action
 
     public function get(Request $request, Response $response, $args)
     {
+        if ($this->tokenWithError($response, $request)){
+            return $response->withStatus(401);
+        }
+        $userRole = $request->getHeader('userRole')[0];
+
+        if ($userRole != AuthorizeMiddleware::GLOBAL_ADMIN){
+            return $response->withStatus(403);
+        }
+
         $idOrganization = (int)$args['id'];
         $idLocalAdmin = (int)$args['idLocalAdmin'];
 
@@ -55,6 +91,15 @@ class LocalAdminAction extends Action
 
     public function getAll(Request $request, Response $response, $args)
     {
+        if ($this->tokenWithError($response, $request)){
+            return $response->withStatus(401);
+        }
+        $userRole = $request->getHeader('userRole')[0];
+
+        if ($userRole != AuthorizeMiddleware::GLOBAL_ADMIN){
+            return $response->withStatus(403);
+        }
+
         $idOrganization = (int)$args['id'];
         $localAdmins = $this->localAdminService->getAll($idOrganization);
 
@@ -63,5 +108,15 @@ class LocalAdminAction extends Action
         }
 
         return $this->respond(200, $localAdmins, $response);
+    }
+
+    public function update(Request $request, Response $response, $args)
+    {
+
+    }
+
+    public function add(Request $request, Response $response, $args)
+    {
+
     }
 }
